requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dash==2.16.1
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_uploader==0.7.0a1
dill==0.3.8
diskcache==5.6.3
dnspython==2.6.1
Flask==3.0.2
Flask-Login==0.6.3
h5py==3.10.0
idna==3.6
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multiprocess==0.70.16
nest-asyncio==1.6.0
numpy==1.26.4
packaging==21.3
pandas==2.2.1
plotly==5.19.0
psutil==5.9.8
pymongo==4.6.2
pyparsing==3.1.2
python-dotenv==1.0.0
PySmbClient==0.1.5
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
retrying==1.3.4
six==1.16.0
tenacity==8.2.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
zipp==3.17.0
